Support in-memory database
