+++ Hippo/appsettings.Development.json
@@ -11,6 +11,7 @@
     "Issuer": "localhost",
     "Audience": "hippos.rocks"
   },
+  "InMemoryDB": "false",
   "ConnectionStrings": {
     "Hippo": "Data Source=hippo.db;Cache=Shared"
   }
